+++ studies/example_study/scenarios/example_im.csv
@@ -1,3 +1,4 @@
-scenario_name,op_module,rq_file,op_fleet_composition,broker_type,pt_operator_type,gtfs_name,pt_simulation_start_date,rq_type,op_rh_reservation_max_routes,evaluation_method,sim_env,user_max_decision_time
-example_im_ptbroker,PoolingIRSOnly,example_100_intermodal.csv,default_vehtype:5,PTBroker,PTControlBasic,example_gtfs,20250101,BasicIntermodalRequest,5,intermodal_evaluation,ImmediateDecisionsSimulation,0
-example_im_ptbroker_lmwt30,PoolingIRSOnly,example_100_intermodal_lmwt30.csv,default_vehtype:5,PTBroker,PTControlBasic,example_gtfs,20250101,BasicIntermodalRequest,5,intermodal_evaluation,ImmediateDecisionsSimulation,0
+scenario_name,op_module,rq_file,op_fleet_composition,broker_type,broker_maas_detour_time_factor,pt_operator_type,gtfs_name,pt_simulation_start_date,rq_type,op_rh_reservation_max_routes,evaluation_method,sim_env,user_max_decision_time
+example_im_ptbroker,PoolingIRSOnly,example_100_intermodal.csv,default_vehtype:5,PTBroker,,PTControlBasic,example_gtfs,20250101,BasicIntermodalRequest,5,intermodal_evaluation,ImmediateDecisionsSimulation,0
+example_im_ptbroker_lmwt30,PoolingIRSOnly,example_100_intermodal_lmwt30.csv,default_vehtype:5,PTBroker,,PTControlBasic,example_gtfs,20250101,BasicIntermodalRequest,5,intermodal_evaluation,ImmediateDecisionsSimulation,0
+example_im_ptbrokerEI_mdtf30,PoolingIRSOnly,example_100_intermodal.csv,default_vehtype:5,PTBrokerEI,30,PTControlBasic,example_gtfs,20250101,BasicIntermodalRequest,5,intermodal_evaluation,ImmediateDecisionsSimulation,0
